+++ src/main/java/com/example/howmuch/config/security/Token.java
@@ -1,4 +1,4 @@
-package com.example.howmuch.contant;
+package com.example.howmuch.config.security;
 
 import lombok.*;
