+++ src/main/java/com/example/howmuch/constant/MyType.java
@@ -1,4 +1,4 @@
-package com.example.howmuch.contant;
+package com.example.howmuch.constant;
 
 
 import lombok.Getter;
