+++ src/main/java/com/example/howmuch/domain/entity/AcEvent.java
@@ -1,7 +1,7 @@
 package com.example.howmuch.domain.entity;
 
-import com.example.howmuch.contant.AcType;
-import com.example.howmuch.contant.EventCategory;
+import com.example.howmuch.constant.AcType;
+import com.example.howmuch.constant.EventCategory;
 import com.example.howmuch.domain.BaseTimeEntity;
 import com.example.howmuch.dto.event.GetAllAcEventsResponse;
 import lombok.*;
