+++ src/main/java/com/example/howmuch/domain/entity/MyEvent.java
@@ -1,7 +1,7 @@
 package com.example.howmuch.domain.entity;
 
-import com.example.howmuch.contant.EventCategory;
-import com.example.howmuch.contant.MyType;
+import com.example.howmuch.constant.EventCategory;
+import com.example.howmuch.constant.MyType;
 import com.example.howmuch.domain.BaseTimeEntity;
 import com.example.howmuch.dto.event.GetAllMyEventsResponse;
 import lombok.*;
