+++ src/main/java/com/example/howmuch/domain/repository/MyEventRepository.java
@@ -1,7 +1,7 @@
 package com.example.howmuch.domain.repository;
 
-import com.example.howmuch.contant.EventCategory;
-import com.example.howmuch.contant.MyType;
+import com.example.howmuch.constant.EventCategory;
+import com.example.howmuch.constant.MyType;
 import com.example.howmuch.domain.entity.MyEvent;
 import com.example.howmuch.domain.entity.User;
 import org.springframework.data.domain.Sort;
